requirements.txt
discord.py>=2.0.0
scikit-learn>=1.0
pandas
numpy
joblib
python-dotenv
